Move the TestLatestStable test to issue-test.yml with  another tag 'yb_version_latest_stable'
