+++ .github/workflows/issue-tests.yml
@@ -48,3 +48,9 @@ jobs:
       run: |
         cd yb-voyager
         go test -v ./... -tags 'issues_integration'
+
+    - name: Test Latest Stable YB Version
+      if: ${{ matrix.version == '2024.1.3.1-b8' }}
+      run: |
+        cd yb-voyager
+        go test -v ./... -tags 'yb_version_latest_stable'
