+++ yb-voyager/src/ybversion/yb_version_test.go
@@ -19,10 +19,6 @@ limitations under the License.
 package ybversion
 
 import (
-	"encoding/json"
-	"io"
-	"net/http"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -76,6 +72,7 @@ func TestPreviewReleaseType(t *testing.T) {
 		ybVersion, _ := NewYBVersion(v)
 		assert.Equal(t, PREVIEW, ybVersion.ReleaseType())
 	}
+	assert.True(t, V2024_2_0_0.GreaterThanOrEqual(V2_25_0_0))
 }
 
 func TestStableOldReleaseType(t *testing.T) {
@@ -86,41 +83,5 @@ func TestStableOldReleaseType(t *testing.T) {
 	for _, v := range stableOldVersionStrings {
 		ybVersion, _ := NewYBVersion(v)
 		assert.Equal(t, STABLE_OLD, ybVersion.ReleaseType())
-	}
-}
-
-func TestLatestStable(t *testing.T) {
-	type Release struct {
-		Name string `json:"name"`
-	}
-
-	url := "https://api.github.com/repos/yugabyte/yugabyte-db/releases"
-	response, err := http.Get(url)
-	assert.NoErrorf(t, err, "could not access URL:%q", url)
-	defer response.Body.Close()
-	if response.StatusCode == 403 {
-		t.Skip("skipping test; rate limit exceeded")
-	}
-	assert.Equal(t, 200, response.StatusCode)
-
-	body, err := io.ReadAll(response.Body)
-	assert.NoError(t, err, "could not read contents of response %q")
-	var releases []Release
-	err = json.Unmarshal(body, &releases)
-	assert.NoErrorf(t, err, "could not unmarshal response %q", string(body))
-	assert.NotEmpty(t, releases, "no releases found")
-	for _, r := range releases {
-		// sample -  v2.20.7.1 (Released October 16, 2024)
-		releaseName := r.Name
-		releaseName = strings.Split(releaseName, " ")[0]
-		if releaseName[0] == 'v' {
-			releaseName = releaseName[1:]
-		}
-		releaseName = strings.Trim(releaseName, " ")
-		rVersion, err := NewYBVersion(releaseName)
-		assert.NoErrorf(t, err, "could not create version %q", releaseName)
-		if rVersion.ReleaseType() == STABLE {
-			assert.True(t, LatestStable.GreaterThanOrEqual(rVersion), "%s is not greater than %s", LatestStable, rVersion)
-		}
 	}
 }
